refactor: fixing directories on imports
- the reason for this repair is because I am having problems during execution between docker, ide and cli;
- both display different errors and I had to fix it to make it work between the three.
