+++ moneymind/src/insert/insert_consolidated_group_data.py
@@ -1,5 +1,5 @@
-from src.get.get_consolidated_group import get_consolidated_group
-from src.mysql_initializer import establishing_mysql_connection
+from get.get_consolidated_group import get_consolidated_group
+from mysql_initializer import establishing_mysql_connection
 
 connection = establishing_mysql_connection()
 cursor = connection.cursor()
